# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.CreateNotificationRequest:
    properties:
      body:
        type: string
      channel:
        allOf:
        - $ref: '#/definitions/models.NotificationChannel'
        description: Channel is the primary notification channel (sms, email, push,
          in_app, webhook)
      idempotencyKey:
        type: string
      locale:
        type: string
      metadata:
        additionalProperties: true
        type: object
      priority:
        allOf:
        - $ref: '#/definitions/models.NotificationPriority'
        description: Priority defaults to normal
      providerId:
        type: string
      recipient:
        allOf:
        - $ref: '#/definitions/dto.NotificationRecipientRequest'
        description: Recipient is the canonical nested recipient object
      recipientEmail:
        type: string
      recipientId:
        type: string
      recipientPhone:
        type: string
      scheduledAt:
        type: string
      subject:
        type: string
      templateId:
        type: string
      templateKey:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.NotificationType'
        description: Type is a backward-compatible alias for channel
      userId:
        description: Legacy flat recipient fields (backward compat)
        type: string
    type: object
...
  dto.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
      requestId:
        type: string
    type: object
...
  dto.PaginatedNotificationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.NotificationResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: 127.0.0.1:9002
info:
  contact:
    email: support@minisource.com
    name: API Support
  description: Notification Service for Minisource - Handles Email, SMS, Push, and
    In-App notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Notifier Service API
  version: "2.0"
paths:
...
  /admin/notifications:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all notifications with full filters
        (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by tenant ID
        in: query
        name: tenantId
        type: string
      - description: Search in recipient/subject/body
        in: query
        name: search
        type: string
      - description: Start date (ISO8601)
        in: query
        name: from
        type: string
      - description: End date (ISO8601)
        in: query
        name: to
        type: string
      - default: createdAt
        description: Sort field
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort direction
        in: query
        name: sortDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedNotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List all notifications'
      tags:
      - Admin Notifications
    post:
      consumes:
      - application/json
      description: 'Create a new notification (admin only).\nCanonical format: {\"channel\":\"sms\",\"recipient\":{\"phone\":\"+989...\"},\"body\":\"...\"}'
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Create notification'
      tags:
      - Admin Notifications
...
  /admin/providers:
    get:
      consumes:
      - application/json
      description: Retrieve all configured notification providers from providers table
      parameters:
      - description: Filter by channel (sms, email, push, webhook, in_app)
        in: query
        name: channel
        type: string
      - description: Filter by status (active, inactive, disabled, error)
        in: query
        name: status
        type: string
      - description: Filter by provider type (e.g., kavenegar, smtp)
        in: query
        name: providerType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProviderResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create a new notification provider
      parameters:
      - description: Provider data
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create provider
      tags:
      - Providers
...
  /admin/reminders:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all reminders with filters (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Start date range (ISO8601)
        in: query
        name: scheduledFrom
        type: string
      - description: End date range (ISO8601)
        in: query
        name: scheduledTo
        type: string
      - description: Search in subject/templateKey
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List all reminders'
      tags:
      - Admin Reminders
    post:
      consumes:
      - application/json
      description: Create a scheduled reminder for any user (admin only)
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Reminder data
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Create reminder'
      tags:
      - Admin Reminders
...
  /admin/templates:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of templates (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Filter by locale
        in: query
        name: locale
        type: string
      - description: Search in name/key/subject
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List all templates'
      tags:
      - Admin Templates
    post:
      consumes:
      - application/json
      description: Create a new notification template (admin only)
      parameters:
      - description: Template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Create template'
      tags:
      - Admin Templates
...
  /me/notifications:
    get:
      consumes:
      - application/json
      description: Retrieve paginated notifications for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Start date (ISO8601)
        in: query
        name: from
        type: string
      - description: End date (ISO8601)
        in: query
        name: to
        type: string
      - description: Search in subject/body
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedNotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my notifications
      tags:
      - User Notifications
...
  /me/preferences:
    get:
      consumes:
      - application/json
      description: Retrieve notification preferences for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PreferenceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my notification preferences
      tags:
      - User Preferences
    put:
      consumes:
      - application/json
      description: Update notification preferences for the authenticated user
      parameters:
      - description: Preference data
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PreferenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my notification preferences
      tags:
      - User Preferences
...
  /me/reminders:
    get:
      consumes:
      - application/json
      description: Retrieve paginated reminders for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Start date range (ISO8601)
        in: query
        name: scheduledFrom
        type: string
      - description: End date range (ISO8601)
        in: query
        name: scheduledTo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my reminders
      tags:
      - User Reminders
    post:
      consumes:
      - application/json
      description: Create a new scheduled reminder for the authenticated user
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Reminder data (userId is ignored, taken from JWT)
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a reminder for myself
      tags:
      - User Reminders
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all notifications with filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by type/channel
        in: query
        name: type
        type: string
      - description: Filter by channel (alias for type)
        in: query
        name: channel
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Search in recipient/subject/body
        in: query
        name: search
        type: string
      - description: Start date (ISO8601)
        in: query
        name: from
        type: string
      - description: End date (ISO8601)
        in: query
        name: to
        type: string
      - default: createdAt
        description: Sort field (createdAt, updatedAt, sentAt, priority, status)
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort direction (asc, desc)
        in: query
        name: sortDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedNotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all notifications (admin)
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Create a new notification. Supports Idempotency-Key header or body
        idempotencyKey field to prevent duplicates.
      parameters:
      - description: Idempotency key to prevent duplicate sends
        in: header
        name: Idempotency-Key
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create notification
      tags:
      - Notifications
...
  /templates:
    get:
      consumes:
      - application/json
      description: Retrieve all notification templates with pagination and optional
        filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by channel/type
        in: query
        name: channel
        type: string
      - description: Filter by locale
        in: query
        name: locale
        type: string
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all templates
      tags:
      - Templates
    post:
      consumes:
      - application/json
      description: Create a new notification template
      parameters:
      - description: Template data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create notification template
      tags:
      - Templates
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"